backend/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Base de datos async
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Validación y settings
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.20

# Autenticación Azure AD / JWT
msal==1.31.1
python-jose[cryptography]==3.3.0
httpx==0.28.1

# Excel / CSV
openpyxl==3.1.5
pandas==2.2.3

# PDF (etiquetas)
reportlab==4.2.5

# Utilidades
python-dotenv==1.0.1
python-slugify==8.0.4
structlog==24.4.0

# Tests
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1
